+++ app/Http/Controllers/Admin/ExpenseCategoryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\ExpenseCategory;
 use Illuminate\Http\Request;
 use Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class ExpenseCategoryController extends Controller
 {
@@ -13,12 +14,14 @@ class ExpenseCategoryController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('expensecategory_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $categories = ExpenseCategory::all();
         return view('admin.expense_categories.index', compact('categories'));
     }
 
     public function create()
     {
+        abort_if(Gate::denies('expensecategory_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('admin.expense_categories.create');
     }
 
@@ -37,6 +40,7 @@ class ExpenseCategoryController extends Controller
 
     public function edit(ExpenseCategory $expenseCategory)
     {
+        abort_if(Gate::denies('expensecategory_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('admin.expense_categories.edit', compact('expenseCategory'));
     }
 
@@ -55,11 +59,13 @@ class ExpenseCategoryController extends Controller
 
     public function show(ExpenseCategory $expenseCategory)
     {
+        abort_if(Gate::denies('expensecategory_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('admin.expense_categories.show', compact('expenseCategory'));
     }
 
     public function destroy(ExpenseCategory $expenseCategory)
     {
+        abort_if(Gate::denies('expensecategory_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $expenseCategory->delete();
         return redirect()->route('admin.expense_categories.index')->with('success', 'Category deleted successfully');
     }
